Variable correction

+++ src/BonusMilesService.java
@@ -1,8 +1,7 @@
 public class BonusMilesService {
     public int calculate(int cost) {
         int bonus_rub = 20;
-        int miles = (int) (cost) / bonus_rub;
-        //       System.out.println(miles);
+        int miles = cost / bonus_rub;
         return miles;
     }
 }
